fix(#2458 review): use u16 (not u32) for proto index to avoid Insn growth
Widening the proto-index field to u32 in MakeFunction/MakeClass/MakeClassMeta
grew size_of::<Insn>() from 24 to 28 bytes. Insn is the dispatched type in
the VM's run_bytecode_inner loop, so the larger instruction word costs a
reproducible ~2-4% on a call-heavy tight loop (hyperfine, /tmp binaries:
u32 PR 339.9ms vs base 334.7ms vs this u16 fix 328.9ms).

u16 caps the per-scope proto count at 65535, which both fully resolves #2444
(no real program defines 65k sibling functions/classes/lambdas/comprehensions
in one scope) and keeps Insn at 24 bytes (measured: u16 -> 24, u32 -> 28).
It also matches the existing bytecode pool widths: name_idx and the consts
index are already u16, and their overflow guards already say "max 65535".

Updated the three Insn variants, the build_def_proto/build_class_proto return
types, the exec_make_* proto_idx params, the optimizer inliner plumbing, and
the four compiler overflow guards (now u16::MAX, wording "max 65535").

+++ crates/pyrust/src/bytecode.rs
@@ -469,7 +469,7 @@ pub enum Insn {
     /// `annots_n == 0` means no annotations; `annots_base` is ignored in that case.
     /// The annotation names (parallel to the register values) are stored in
     /// `FnProto::annotation_keys`.
-    MakeFunction(Reg, u32, Reg, u8, Reg, u8),
+    MakeFunction(Reg, u16, Reg, u8, Reg, u8),
     /// R[dst] = load_module(names[name_idx])
     ImportModule(Reg, u16),
 
@@ -523,7 +523,7 @@ pub enum Insn {
     /// R[dst] = create class(fn_protos[proto_idx], bases R[bases_base..+bases_n], name=names[name_idx])
     /// PEP 487: kwarg_n keyword arg values are in R[kwarg_base..kwarg_base+kwarg_n];
     /// names come from fn_protos[proto_idx].class_kwarg_names.
-    MakeClass(Reg, u32, Reg, u8, u16, Reg, u8),
+    MakeClass(Reg, u16, Reg, u8, u16, Reg, u8),
     /// R[dst] = metaclass-driven class creation, where the metaclass value is in
     /// R[meta_reg].  Same layout as `MakeClass` plus a trailing `meta_reg`.
     /// Unlike `MakeClass`, this calls `metaclass.__prepare__(name, bases, **kw)`
@@ -533,7 +533,7 @@ pub enum Insn {
     /// (`type.__new__`) rather than in `MakeClass` (issues #2128/#2130).
     /// Tuple: (dst, proto_idx, bases_base, bases_n, name_idx, kwarg_base,
     /// kwarg_n, meta_reg).
-    MakeClassMeta(Reg, u32, Reg, u8, u16, Reg, u8, Reg),
+    MakeClassMeta(Reg, u16, Reg, u8, u16, Reg, u8, Reg),
     /// R[dst] = TypeVar(name=consts[name_idx])
     /// PEP 695: construct an (initially unbounded) `TypeVar` object for a generic
     /// type parameter — `__bound__` is `None` and `__constraints__` is `()`.
